requirements.txt
Flask
PyYAML
Flask-SQLAlchemy
Flask-Migrate
paramiko
google-generativeai
openai
requests
PyGithub
Flask-CORS
Authlib
Flask-Login
APScheduler
croniter